agent/requirements.txt
smolagents==1.13.0
requests
duckduckgo_search
pandas
markdownify
fastapi
uvicorn
pydantic
python-dotenv
pytz
pyyaml